Crate pallet_erc20_xcm_bridge

Pallet that allow to transact erc20 tokens trought xcm directly.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

AssetTrapWrapper

Morph a given DropAssets implementation into one which filter out erc20 assets.

XcmExecuteFilterWrapper

Morph a given “xcm execution” filter implementation into one which filter out some XCM instructions if they could manipulate erc20 assets. If your runtime allows arbitrary xcm messages to be executed locally you should use this

XcmExecutorWrapper

Xcm executor wrapper that inject xcm holding extension “XcmHoldingErc20sOrigins”