pallet_erc20_xcm_bridge::pallet
Trait Config

pub trait Config: Config + Config {
    type AccountIdConverter: ConvertLocation<H160>;
    type Erc20MultilocationPrefix: Get<Location>;
    type Erc20TransferGasLimit: Get<u64>;
    type EvmRunner: Runner<Self>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type AccountIdConverter: ConvertLocation<H160>

type Erc20MultilocationPrefix: Get<Location>

type Erc20TransferGasLimit: Get<u64>

type EvmRunner: Runner<Self>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.